Move profile.hpp
