+++ tools/include/blanc/compiler_options.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-#include "internal/profile.hpp"
+#include "./profile.hpp"
